Accept that test folder can already exists

+++ cli.js
@@ -6,12 +6,15 @@ const struct = require('./templates/initial-file-structure');
 
 function createDir(dirPath) {
   const testPath = dirPath.replace(/src/, 'test');
+  if (fs.existsSync(testPath)) return;
 
   fs.mkdirSync(testPath);
 }
 
 function createFile(filePath) {
   const testPath = filePath.replace(/src/, 'test').replace(/.js/, '.spec.js');
+  if (fs.existsSync(testPath)) return;
+
   const fileName = path.basename(testPath, '.spec.js');
 
   fs.writeFileSync(testPath, struct(fileName), { encoding: 'utf8' });
@@ -33,10 +36,10 @@ function recursivelyReadDir(dirPath) {
 }
 
 function generate() {
-  if (fs.existsSync(`${__dirname}/test`)) 
-    throw new Error('Test directory already exists');
+  if (!fs.existsSync(`${__dirname}/test`)) {
+    fs.mkdirSync(`${__dirname}/test`);
+  } 
 
-  fs.mkdirSync(`${__dirname}/test`);
   recursivelyReadDir(path.join(__dirname, 'src'), '');
 }
